Teleop works
Removed useless confusing files that no one wants at all and got tank
and arcade teleop fixed
