--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TankArcadeDrive/ArcadeDrive.java
@@ -1,55 +1,0 @@
-package org.firstinspires.ftc.teamcode.TankArcadeDrive;
-
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.util.Range;
-
-import org.firstinspires.ftc.teamcode.TankArcadeDrive.DriveOPBase;
-
-/**
- * Created by Eric Golde on 9/20/2016.
- */
-
-//simple arcade drive. Use joystick #1
-@TeleOp(name="ArcadeDrive", group="Drive")
-@Disabled
-public class ArcadeDrive extends DriveOPBase {
-
-    DcMotor left;
-    DcMotor right;
-
-    @Override
-    public void init() {
-        left = hardwareMap.dcMotor.get("left");
-        right = hardwareMap.dcMotor.get("right");
-        right.setDirection(DcMotorSimple.Direction.REVERSE);
-    }
-
-    @Override
-    public void start(){
-
-    }
-
-    @Override
-    public void loop() {
-
-        float xValue = gamepad1.left_stick_x;
-        float yValue = -gamepad1.left_stick_y;
-
-        float leftPower = yValue + xValue;
-        float rightPower = yValue - xValue;
-
-        leftPower = Range.clip(leftPower, -DriveOPBase.motorSpeed.getValueFloat(), DriveOPBase.motorSpeed.getValueFloat());
-        rightPower = Range.clip(rightPower, -DriveOPBase.motorSpeed.getValueFloat(), DriveOPBase.motorSpeed.getValueFloat());
-
-        left.setPower(leftPower);
-        right.setPower(rightPower);
-    }
-
-    @Override
-    public void stop(){
-
-    }
-}
