+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TankChassis/TankDriveTeleop.java
@@ -3,8 +3,6 @@ package org.firstinspires.ftc.teamcode.TankChassis;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.Range;
 
-import org.firstinspires.ftc.teamcode.TankArcadeDrive.DriveOPBase;
-
 /**
  * Created by Eric Golde on 10/8/2016.
  */
@@ -26,8 +24,8 @@ public class TankDriveTeleop extends TankChassisBase {
         float rightValue = -gamepad1.right_stick_y;
 
         //clip the variables to be what the config says
-        leftValue = Range.clip(leftValue, TankChassisConfig.motorSpeed.getValueFloat(), TankChassisConfig.motorSpeed.getValueFloat());
-        rightValue = Range.clip(rightValue, TankChassisConfig.motorSpeed.getValueFloat(), TankChassisConfig.motorSpeed.getValueFloat());
+        leftValue = Range.clip(leftValue, -TankChassisConfig.motorSpeed.getValueFloat(), TankChassisConfig.motorSpeed.getValueFloat());
+        rightValue = Range.clip(rightValue, -TankChassisConfig.motorSpeed.getValueFloat(), TankChassisConfig.motorSpeed.getValueFloat());
 
         left.setPower(leftValue);
         right.setPower(rightValue);
